Update the manual interactive-browser-credential test
- use the template project from `npx create-react-app --template
typescript ...`

- use `react-app-rewired` to customize webpack configurations

- move our testing code into App.tsx
