+++ sdk/identity/identity/test/manual/interactive-browser-credential/README.md
@@ -14,9 +14,8 @@ npm install
 npm start
 ```
 
-Webpack will compile the code and then host it in a local server at
-`http://localhost:8080`.  See the section below about CORS before attempting to
-navigate to this URL.
+React tooling will compile the code and then host it in a local server at
+`http://localhost:3000`.
 
 You will need to configure an AAD App Registration as follows:
 
@@ -25,7 +24,7 @@ You will need to configure an AAD App Registration as follows:
 - Go to the App Registration section in the AAD page.
 - Click on the app that you want to use to authenticate.
 - Go to the Authentication tab of your AAD application.
-- Click on `+ Add a platform`, select `Single-page application`, enter `http://localhost:8080` as the redirect URI, then make sure to include implicit grant for "Access tokens" and "ID tokens".
+- Click on `+ Add a platform`, select `Single-page application`, enter `http://localhost:3000` as the redirect URI, then make sure to include implicit grant for "Access tokens" and "ID tokens".
 - Go to the `API permissions` tab of your AAD application. Click on `Add a permission`, then go to `APIs my organization uses` and search for `Microsoft.ServiceBus`, then add this permission.
 - Keep in mind that if you belong to an organization, other restrictions based on the organization configurations might prevent you from authenticating. If these steps don't end up being effective, try again on a personal account.
 
@@ -35,7 +34,7 @@ Grant access to this AAD application to your Service Bus by:
 - Either in the "Access policies" section of the creation form, or by going to the Service Bus namespace's "Access policies" page, click con `+ Add Access Policy`, select all permissions, then select your AAD application as the "Service Bus Data Owner", then click "Add", then click "Save" if applicable.
 - Then, in your Service Bus namespace, create a queue named `queue-identity-test`.
 
-With the AAD application and the Service Bus namespace configured, make sure `npm start` is running, then go to `http://localhost:8080`, then enter the Tenant ID, the Client ID of the AAD application and the Service Bus Endpoint, then go through the available options to configure the scenario in which you want the authentication to run, then click on the `Send Message` button. Information from the message sent will appear at the bottom.
+With the AAD application and the Service Bus namespace configured, make sure `npm start` is running, then go to `http://localhost:3000`, then enter the Tenant ID, the Client ID of the AAD application and the Service Bus Endpoint, then go through the available options to configure the scenario in which you want the authentication to run, then click on the `Send Message` button. Information from the message sent will appear at the bottom.
 
 If something unexpected happens, make sure to open the console tab in the browser. The application will be logging as many things as they seemed relevant for debugging.
