--- sdk/identity/identity/test/manual/interactive-browser-credential/index.html
@@ -1,8 +1,0 @@
-<html>
-
-<body>
-  <div id="app"></div>
-</body>
-<script type="text/javascript" src="dist/index.js"></script>
-
-</html>
