+++ sdk/identity/identity/test/manual/interactive-browser-credential/package.json
@@ -1,37 +1,49 @@
 {
   "name": "identity-browser-manual-tests",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Manual tests for Azure SDK browser use scenarios",
-  "main": "dist/bundle.js",
-  "scripts": {
-    "build": "webpack --config webpack.config.js",
-    "start": "webpack-dev-server",
-    "check-format": "prettier --list-different --config ../../../../../.prettierrc.json --ignore-path ../../../../../.prettierignore \"src/**/*.tsx\"",
-    "format": "prettier --write --config ../../../../../.prettierrc.json --ignore-path ../../../../../.prettierignore \"src/**/*.tsx\""
-  },
-  "author": "Microsoft Corporation",
-  "license": "MIT",
+  "private": true,
   "dependencies": {
-    "@azure/identity": "^2.0.1",
-    "@azure/service-bus": "^7.0.3",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "tslib": "2.4.1"
+    "@azure/identity": "^3.1.3",
+    "@azure/identity-cache-persistence": "^1.0.0",
+    "@azure/service-bus": "^7.8.1",
+    "@types/node": "^16.18.16",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "react-app-rewired start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   },
   "devDependencies": {
-    "@azure/identity-cache-persistence": "^1.0.0",
-    "@types/express": "^4.16.0",
-    "@types/node": "14.0.0",
-    "@types/react": "^16.8.24",
-    "@types/react-dom": "^16.8.5",
-    "@types/webpack": "^4.4.13",
-    "@types/webpack-dev-middleware": "^2.0.2",
-    "dotenv": "^16.0.0",
-    "prettier": "^2.5.1",
-    "ts-loader": "^8.0.0",
-    "typescript": "~4.8.0",
-    "webpack": "^4.0.0",
-    "webpack-cli": "^3.0.0",
-    "webpack-dev-server": "^3.7.2"
+    "buffer": "^6.0.3",
+    "os-browserify": "^0.3.0",
+    "path-browserify": "^1.0.1",
+    "process": "^0.11.10",
+    "react-app-rewired": "^2.2.1"
   }
 }
