+++ sdk/identity/identity/test/manual/interactive-browser-credential/public/robots.txt
@@ -1,0 +1,3 @@
+# https://www.robotstxt.org/robotstxt.html
+User-agent: *
+Disallow:
