+++ sdk/identity/identity/test/manual/interactive-browser-credential/src/react-app-env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="react-scripts" />
