--- sdk/identity/identity/test/manual/interactive-browser-credential/webpack.config.js
@@ -1,29 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  entry: './src/index.tsx',
-  output: {
-    filename: 'index.js',
-    path: path.resolve(__dirname, 'dist')
-  },
-  module: {
-    rules: [
-      {
-        test: /\.tsx?$/,
-        use: 'ts-loader',
-        exclude: /node_modules/
-      }
-    ]
-  },
-  resolve: {
-    extensions: [ '.tsx', '.ts', '.js' ],
-    aliasFields: ['browser']
-  },
-  mode: 'development',
-  devServer: {
-    contentBase: '.',
-  },
-  optimization: {
-    usedExports: true
-  }
-};
